npm: remove the "install" phase from the package.json
npm is actually smart enough to invoke node-waf when it sees a `wscript`
file in the package root (for a very long time), or node-gyp when it sees
a `binding.gyp` file (as of npm v1.1.8 and above).

It runs them with the default commands so we have to depend on the default
build location that node-waf/node-gyp builds the module into, which would
be: `./build/Release/pty.node`, so updating lib/pty.js to use that path
as well, and updating the Makefile to not bother doing that anymore.