Don't resume socket on initialization
The standard scenario is that consumers will add a data event handler which
will resume the stream. This ensures that no data events have gone lost if
there was a delay between the consumer initializing the pty and attaching a
data handler.

To get the exact previous behavior just call pty.resume() after initialization.

Fixes #99