Make WindowsTerminal throw exceptions again
The previous patch changed WindowsTerminal to make it report errors
(e.g. trying to execute a non-existent file) asynchronously (through a
callback that the user would register after creating the Terminal
object).  This turns out to be too complicated, because we end up with
an half-backed object, with some fields that are undefined, but
shouldn't.

Instead, revert to the previous behavior of throwing an exception on
Windows when executing fails, it works well enough.  To work well
cross-platform, the user code will need to both catch exceptions and
register an exec callback.  This is not ideal, but it's enough for now.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>