Remove dynamic loading method `loadNative`
Use of this dynamic loading method is convenient, but makes it difficult for bundlers like webpack to locate files. With a bit of extra bloat, users of webpack will not need to treat the node-pty project as an 'external', and will be able to bundle it just like any other node module.